Memory/QMD: harden no-results parsing
